+++ src/index.ts
@@ -37,9 +37,9 @@ import * as RefundActionFactory from './action/trade/refund';
 import * as GiveActionFactory from './action/transfer/give';
 import * as GivePointAwardActionFactory from './action/transfer/give/pointAward';
 import * as MoneyTransferActionFactory from './action/transfer/moneyTransfer';
+import * as ReturnInvoiceActionFactory from './action/transfer/return/invoice';
 import * as ReturnMoneyTransferActionFactory from './action/transfer/return/moneyTransfer';
 import * as ReturnOrderActionFactory from './action/transfer/return/order';
-import * as ReturnPaymentMethodActionFactory from './action/transfer/return/paymentMethod';
 import * as ReturnPointAwardActionFactory from './action/transfer/return/pointAward';
 import * as ReturnReserveTransactionActionFactory from './action/transfer/return/reserveTransaction';
 import * as SendEmailMessageActionFactory from './action/transfer/send/message/email';
@@ -198,7 +198,6 @@ import * as RegisterServiceAssetTransactionFactory from './assetTransaction/regi
 import * as ReserveAssetTransactionFactory from './assetTransaction/reserve';
 import { AssetTransactionType } from './assetTransactionType';
 import { TransactionStatusType } from './transactionStatusType';
-import { TransactionTasksExportationStatus } from './transactionTasksExportationStatus';
 
 import * as TransactionFactory from './transaction';
 import * as MoneyTransferTransactionFactory from './transaction/moneyTransfer';
@@ -309,11 +308,11 @@ export namespace action {
          */
         export namespace returnAction {
             // tslint:disable-next-line:no-shadowed-variable
+            export import invoice = ReturnInvoiceActionFactory;
+            // tslint:disable-next-line:no-shadowed-variable
             export import moneyTransfer = ReturnMoneyTransferActionFactory;
             // tslint:disable-next-line:no-shadowed-variable
             export import order = ReturnOrderActionFactory;
-            // tslint:disable-next-line:no-shadowed-variable
-            export import paymentMethod = ReturnPaymentMethodActionFactory;
             export import pointAward = ReturnPointAwardActionFactory;
             export import reserveTransaction = ReturnReserveTransactionActionFactory;
         }
@@ -825,7 +824,7 @@ export namespace transaction {
 }
 export import transactionType = TransactionType;
 export import transactionStatusType = TransactionStatusType;
-export import transactionTasksExportationStatus = TransactionTasksExportationStatus;
+// export import transactionTasksExportationStatus = TransactionTasksExportationStatus; // discontinue(2024-06-20~)
 export import assetTransactionType = AssetTransactionType;
 export import unitCode = UnitCode;
 export import unitPriceOffer = UnitPriceOfferFactory;